[Component] Modal window implemented
