+++ frontend/package.json
@@ -5,6 +5,7 @@
   "dependencies": {
     "@auth0/auth0-react": "^1.5.0",
     "@material-ui/core": "^4.11.4",
+    "@material-ui/icons": "^4.11.2",
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
