+++ frontend/src/components/ModalWindowContainer.css
@@ -1,0 +1,23 @@
+.modal-content{
+    width: 500px;
+    margin: auto;
+    position: absolute;
+    top:50%;
+    left:50%;
+    transform: translate(-50%, -50%)
+}
+
+.modal-content-container{
+    position: relative;
+    height:100vh
+}
+.close-btn{
+    text-align: center;
+    margin-bottom: 10px;
+}
+.close-btn:hover{
+    cursor: pointer
+}
+.close-icon{
+    display: inline-block;
+}
