+++ frontend/src/components/create-game/CreateGameButtonContainer.js
@@ -3,7 +3,7 @@ import CommonButton from '../CommonButton'
 const CreateGameButtonContainer = ()=>{
     return (
         <div style={{textAlign: "right", paddingTop: "10px", paddingBottom: "10px"}}>
-            <CommonButton name="Create"/>
+            <CommonButton name="Create" color="primary" />
         </div>
     )
 }
