--- frontend/src/pages/TestPage.css
@@ -1,4 +1,0 @@
-.container{
-    width: 300px;
-    margin: auto;
-}
